requirements.txt
flask==3.0.0
flask-cors==4.0.0
chromadb==0.4.18
anthropic==0.8.1
python-dotenv==1.0.0
gunicorn==21.2.0